Added logic for 404 Not Found

+++ src/main/java/com/example/bookstore_mongodb/controller/BookController.java
@@ -1,9 +1,13 @@
 package com.example.bookstore_mongodb.controller;
 
 import com.example.bookstore_mongodb.entity.Book;
+import com.example.bookstore_mongodb.repository.BookRepository;
 import com.example.bookstore_mongodb.service.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -12,6 +16,8 @@ import java.util.List;
 public class BookController {
     @Autowired
     private BookService bookService;
+    @Autowired
+    private BookRepository bookRepository;
 
     @GetMapping
     public List<Book> getAllBooks() {
@@ -25,7 +31,12 @@ public class BookController {
 
     // Remove a book
     @DeleteMapping("/{id}")
-    public Book deleteBook(@PathVariable String id) {
-        return bookService.deleteBook(id);
+    public ResponseEntity<String> deleteBook(@PathVariable String id) {
+        if (!bookRepository.existsById(id)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
+        bookService.deleteBook(id);
+        return ResponseEntity.ok("Deleted successfully");
     }
 }
